feat: Erro: Byte desempenho; não está claro. Se for um percentual, use Float.

+++ services/static/src/main/java/soccervs/statics/core/entities/Resultados.java
@@ -37,7 +37,7 @@ public class Resultados {
 	private Byte sofridos;
 	
 	@Column
-	private Byte desempenho;
+	private Float desempenho;
 	
 	@ManyToOne @JoinColumn(name = "competicao_id", nullable = false)
 	private Competicoes competicao;
